Handle the case where message.format.version has been set for a topic already
