+++ core/src/test/scala/unit/kafka/cluster/PartitionTest.scala
@@ -32,7 +32,6 @@ import org.apache.kafka.common.protocol.Errors
 import org.apache.kafka.common.record.FileRecords.TimestampAndOffset
 import org.apache.kafka.common.record._
 import org.apache.kafka.common.requests.ListOffsetsRequest
-import org.apache.kafka.common.requests.OffsetsForLeaderEpochResponse.{UNDEFINED_EPOCH, UNDEFINED_EPOCH_OFFSET}
 import org.apache.kafka.common.utils.SystemTime
 import org.apache.kafka.common.{IsolationLevel, TopicPartition, Uuid}
 import org.junit.jupiter.api.Assertions._
@@ -47,7 +46,6 @@ import java.util.Optional
 import java.util.concurrent.{CountDownLatch, Semaphore}
 import kafka.server.epoch.LeaderEpochFileCache
 
-import scala.annotation.nowarn
 import scala.jdk.CollectionConverters._
 
 class PartitionTest extends AbstractPartitionTest {
@@ -158,37 +156,8 @@ class PartitionTest extends AbstractPartitionTest {
     assertEquals(leaderEpoch, epochEndOffset.leaderEpoch)
   }
 
-  @nowarn("cat=deprecation")
-  @Test
-  def testMakeLeaderDoesNotUpdateEpochCacheForOldFormats(): Unit = {
-    val leaderEpoch = 8
-    configRepository.setTopicConfig(topicPartition.topic(),
-      LogConfig.MessageFormatVersionProp, kafka.api.KAFKA_0_10_2_IV0.shortVersion)
-    val log = logManager.getOrCreateLog(topicPartition, topicId = None)
-    log.appendAsLeader(TestUtils.records(List(
-      new SimpleRecord("k1".getBytes, "v1".getBytes),
-      new SimpleRecord("k2".getBytes, "v2".getBytes)),
-      magicValue = RecordVersion.V1.value
-    ), leaderEpoch = 0)
-    log.appendAsLeader(TestUtils.records(List(
-      new SimpleRecord("k3".getBytes, "v3".getBytes),
-      new SimpleRecord("k4".getBytes, "v4".getBytes)),
-      magicValue = RecordVersion.V1.value
-    ), leaderEpoch = 5)
-    assertEquals(4, log.logEndOffset)
-
-    val partition = setupPartitionWithMocks(leaderEpoch = leaderEpoch, isLeader = true)
-    assertEquals(Some(4), partition.leaderLogIfLocal.map(_.logEndOffset))
-    assertEquals(None, log.latestEpoch)
-
-    val epochEndOffset = partition.lastOffsetForLeaderEpoch(currentLeaderEpoch = Optional.of(leaderEpoch),
-      leaderEpoch = leaderEpoch, fetchOnlyFromLeader = true)
-    assertEquals(UNDEFINED_EPOCH_OFFSET, epochEndOffset.endOffset)
-    assertEquals(UNDEFINED_EPOCH, epochEndOffset.leaderEpoch)
-  }
-
-  @Test
   // Verify that partition.removeFutureLocalReplica() and partition.maybeReplaceCurrentWithFutureReplica() can run concurrently
+  @Test
   def testMaybeReplaceCurrentWithFutureReplica(): Unit = {
     val latch = new CountDownLatch(1)
 
@@ -740,40 +709,6 @@ class PartitionTest extends AbstractPartitionTest {
 
     // Now we see None instead of an error for out of range timestamp
     assertEquals(Right(None), fetchOffsetsForTimestamp(100, Some(IsolationLevel.READ_UNCOMMITTED)))
-  }
-
-  private def setupPartitionWithMocks(leaderEpoch: Int,
-                                      isLeader: Boolean): Partition = {
-    partition.createLogIfNotExists(isNew = false, isFutureReplica = false, offsetCheckpoints, None)
-
-    val controllerEpoch = 0
-    val replicas = List[Integer](brokerId, brokerId + 1).asJava
-    val isr = replicas
-
-    if (isLeader) {
-      assertTrue(partition.makeLeader(new LeaderAndIsrPartitionState()
-          .setControllerEpoch(controllerEpoch)
-          .setLeader(brokerId)
-          .setLeaderEpoch(leaderEpoch)
-          .setIsr(isr)
-          .setZkVersion(1)
-          .setReplicas(replicas)
-          .setIsNew(true), offsetCheckpoints, None), "Expected become leader transition to succeed")
-      assertEquals(leaderEpoch, partition.getLeaderEpoch)
-    } else {
-      assertTrue(partition.makeFollower(new LeaderAndIsrPartitionState()
-          .setControllerEpoch(controllerEpoch)
-          .setLeader(brokerId + 1)
-          .setLeaderEpoch(leaderEpoch)
-          .setIsr(isr)
-          .setZkVersion(1)
-          .setReplicas(replicas)
-          .setIsNew(true), offsetCheckpoints, None), "Expected become follower transition to succeed")
-      assertEquals(leaderEpoch, partition.getLeaderEpoch)
-      assertEquals(None, partition.leaderLogIfLocal)
-    }
-
-    partition
   }
 
   @Test
