+++ jmh-benchmarks/src/main/java/org/apache/kafka/jmh/partition/PartitionMakeFollowerBenchmark.java
@@ -17,6 +17,7 @@
 
 package org.apache.kafka.jmh.partition;
 
+import kafka.api.ApiVersion;
 import kafka.api.ApiVersion$;
 import kafka.cluster.DelayedOperations;
 import kafka.cluster.IsrChangeListener;
@@ -111,6 +112,7 @@ public class PartitionMakeFollowerBenchmark {
             10000L,
             1000L,
             60000,
+            ApiVersion.latestVersion(),
             scheduler,
             brokerTopicStats,
             logDirFailureChannel,
